+++ internationalonlineoffer/templates/ioo/header.html
@@ -1,0 +1,22 @@
+{% load static %}
+{% load ga360_data from directory_components %}
+
+<header data-ga-section="header">
+    <div class="atlas-header__global">
+        <div class="atlas-container atlas-container--wide">
+            <img width="283" height="21" alt="Department for International Trade"
+                 src="{% static 'directory_components/images/dit-logo.svg' %}">
+        </div>
+    </div>
+    <div class="atlas-header__main">
+        <div class="atlas-container atlas-container--wide atlas-header__main-inner">
+            {% ga360_data "a" value="logo" %}
+                <a class="atlas-header__logo">
+                    <img src="{% static 'images/great.png' %}" width="217" height="81"
+                         alt="Great Britain and Northern Ireland"/>
+                </a>
+            {% end_ga360_data %}
+
+        </div>
+    </div>
+</header>
