+++ internationalonlineoffer/templates/ioo/index.html
@@ -18,7 +18,7 @@
       </div>
       <div class="grid-row">
         <div class="column">
-          <a href="/international/ioo/sector/" class="atlas-button">Get Started</a>
+          <a class="atlas-button">Start now ></a>
         </div>
       </div>
     </div>
