--- ioo/templates/ioo/header.html
@@ -1,65 +1,0 @@
-{% load static %}
-{% load ga360_data from directory_components %}
-
-<header data-ga-section="header">
-    <div class="atlas-header__global">
-        <div class="atlas-container atlas-container--wide">
-            <img width="283" height="21" alt="Department for International Trade"
-                 src="{% static 'directory_components/images/dit-logo.svg' %}">
-        </div>
-    </div>
-    <div class="atlas-header__main">
-        <div class="atlas-container atlas-container--wide atlas-header__main-inner">
-            {% ga360_data "a" value="logo" %}
-                <a class="atlas-header__logo" href="{{ services_urls.great_international }}">
-                    <img src="{% static 'images/great.png' %}" width="217" height="81"
-                         alt="Great Britain and Northern Ireland"/>
-                </a>
-            {% end_ga360_data %}
-
-            <button id="atlas-nav-toggle" class="atlas-menu-button" aria-controls="atlas-nav" data-reveal-button>
-                Menu
-                <span class="atlas-menu-button__icon"></span>
-            </button>
-
-            <nav id="atlas-nav" class="atlas-nav" aria-label="Main menu">
-                <ul class="atlas-nav__list atlas-list--bare">
-                    {% for node in navigation_tree %}
-                        <li>
-                            <a class="atlas-nav__item{% if node.tier_one_item.name == header_section %} active{% endif %}"
-                               href="{{ node.tier_one_item.url }}">{{ node.tier_one_item.title }}</a>
-                            <ul class="atlas-nav__section atlas-list--bare">
-                                {% for item in node.tier_two_items %}
-                                    <li>
-                                        <a class="atlas-nav__section-item{% if item.name == header_sub_section %} active{% endif %}"
-                                           href="{{ item.url }}">{{ item.title }}</a>
-                                    </li>
-                                {% endfor %}
-                            </ul>
-                        </li>
-                    {% endfor %}
-                </ul>
-
-            </nav>
-        </div>
-    </div>
-
-    {% for node in navigation_tree %}
-        {% if node.tier_one_item.name == header_section and node.tier_two_items %}
-            <div class="atlas-subnav">
-                <div class="atlas-container atlas-container--wide">
-                    <nav class="atlas-subnav" aria-label="Sub-menu">
-                        <ul>
-                            {% for item in node.tier_two_items %}
-                                <li>
-                                    <a class="atlas-subnav__item{% if item.name == header_sub_section %} active{% endif %}"
-                                       href="{{ item.url }}">{{ item.title }}</a>
-                                </li>
-                            {% endfor %}
-                        </ul>
-                    </nav>
-                </div>
-            </div>
-        {% endif %}
-    {% endfor %}
-</header>
