+++ pii-secret-exclude.txt
@@ -59,11 +59,11 @@ domestic/static/styles/contact-selling-online-overseas.css.map
 domestic/static/styles/main.css.map
 config/env/test
 tests/unit/core/test_context_processors.py
-ioo/sass/_tokens.scss
+internationalonlineoffer/sass/_tokens.scss
 config/env/dev
-ioo/templates/ioo/base.html
-ioo/static/fonts/EuclidFlexB-Regular.otf
-ioo/views.py
-ioo/static/fonts/EuclidFlexB-Bold.otf
-ioo/static/fonts/EuclidFlexB-Medium.otf
-ioo/sass/components/_footer.scss
+internationalonlineoffer/static/fonts/EuclidFlexB-Regular.otf
+internationalonlineoffer/views.py
+internationalonlineoffer/static/fonts/EuclidFlexB-Bold.otf
+internationalonlineoffer/static/fonts/EuclidFlexB-Medium.otf
+internationalonlineoffer/sass/components/_footer.scss
+internationalonlineoffer/templates/ioo/base.html
